requirements.txt
kivy==2.3.0
pycryptodome==3.20.0
